Add ipython notebook and code
Exploring and cleaning file asylum_clean.csv.
Implementing a basic Log Reg as an optional baseline model for using
simple input such as the features ’lawyer’ and ‘defensive’. The outcome
is ‘grantraw’.

And

Exploring and cleaning file asylum_clean.csv
Implementing a basic Log Reg as an optional baseline model for using
simple input such as the features ’hearing_loc_code’ and ‘natid. The
outcome is ‘grantraw’